+++ packages/target-tauri/src-tauri/src/state/main_window_channels.rs
@@ -32,6 +32,7 @@ pub enum MainWindowEvents {
     ShowSettingsDialog,
     ShowKeybindingsDialog,
     ResumeFromSleep,
+    ToggleNotifications,
 }
 
 pub(crate) struct InnerMainWindowChannelsState {